Moved shrunk functionality to take place inside of the paddle instead of in the ball.

+++ src/ofApp.cpp
@@ -83,7 +83,7 @@ void ofApp::draw(){
 	}
 	else
 	{
-		logo.draw(ofGetWidth() / 2, ofGetHeight() / 5);
+		logo.draw(static_cast<float>(ofGetWidth()) / 2, static_cast<float>(ofGetHeight()) / 5);
 		startButtonImage.draw(startButtonCoord.x, startButtonCoord.y);
 	}
 }
